refactor: break out the handler for editable content
