+++ spec/JustNotSorrySpec.test.js
@@ -1,6 +1,5 @@
 import { h } from 'preact';
 import JustNotSorry from '../src/components/JustNotSorry.js';
-import { handleContentEditableContentInsert } from '../src/components/JustNotSorry.js';
 import { configure, mount } from 'enzyme';
 import Adapter from 'enzyme-adapter-preact-pure';
 
@@ -366,41 +365,6 @@ describe('JustNotSorry', () => {
 
           expect(spy).toHaveBeenCalledWith(node.getDOMNode());
         });
-      });
-    });
-  });
-
-  describe('#handleContentEditableContentInsert', () => {
-    let mockNode;
-    beforeEach(() => {
-      mockNode = {
-        dispatchEvent: jest.fn(),
-        hasAttribute: () => true,
-      };
-    });
-    describe('when an observed content editable sees a non-text change (such as a line break)', () => {
-      it('should dispatch an input event to trigger checking for warnings', () => {
-        const mockMutation = {
-          type: 'childList',
-          target: mockNode,
-        };
-
-        handleContentEditableContentInsert([mockMutation]);
-
-        expect(mockNode.dispatchEvent).toHaveBeenCalledWith(expect.any(Event));
-      });
-    });
-
-    describe('when an observed content editable sees a text change', () => {
-      it('should NOT dispatch an extra input event', () => {
-        const mockMutation = {
-          type: 'characterData',
-          target: mockNode,
-        };
-
-        handleContentEditableContentInsert([mockMutation]);
-
-        expect(mockNode.dispatchEvent).not.toHaveBeenCalled();
       });
     });
   });
