+++ src/handlers/EditableContent.js
@@ -1,0 +1,13 @@
+const isContentEditableCharacterData = (mutation) =>
+  mutation.type !== 'characterData' &&
+  mutation.target.hasAttribute('contentEditable');
+
+const triggerCheckForWarnings = (mutation) =>
+  mutation.target.dispatchEvent(
+    new Event('input', { bubbles: true, cancelable: true })
+  );
+
+export const handleContentEditableContentInsert = (mutations) =>
+  mutations
+    .filter(isContentEditableCharacterData)
+    .forEach(triggerCheckForWarnings);
